reverted previous change of variable type in MITHNEOS_to_spectrum
line 816, int to float. Previously changed from float to int, but, as it turns out, not all data in MITHNEOS have the same number of columns. Some 200-300 are missing the 4th column for number of frames used as an average to create the data point. Hence, the column with index -1 is the noise as a string, hence throws an exception when casted to int from base 10.

+++ splat/asteroid.py
@@ -813,7 +813,7 @@ def MITHNEOS_to_Spectrum(txtPath):
     for entry in dataList:
         if entry is not '': # ignores empty rows
             removed_spaces = entry.split()
-            if (int(removed_spaces[-1]) != 0): 
+            if (float(removed_spaces[-1]) != 0): 
                 pruned_dataList.append(removed_spaces) # strings, of each value
 
     # pruned_dataList is a list of lists, of which the smaller list are rows of the .txt file
